fix(pwa): address review feedback
